requirements.txt
# Core Django
Django==4.2.25
djangorestframework==3.16.1
dj_database_url

# Database
psycopg2-binary==2.9.11

# Async Tasks
celery==5.5.3
redis==7.0.1
django-celery-beat==2.8.1

# Web Scraping
beautifulsoup4==4.14.2
requests==2.32.5
playwright==1.55.0
fake-useragent==2.2.0

# Document Processing
pdfplumber==0.11.7
python-docx==1.2.0
Pillow==11.3.0
pytesseract==0.3.13

# NLP & AI
google-generativeai==0.8.5
spacy==3.8.7

# Forms & UI
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-widget-tweaks==1.5.0

# Configuration
python-decouple==3.8

# Production
gunicorn==23.0.0
whitenoise==6.11.0

# Utilities
python-slugify==8.0.4
pytz==2025.2

django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
